usage in build script, improved bginfo

+++ build.sh
@@ -1,3 +1,9 @@
+if [ $# -ne 1 ]
+then
+  echo "Usage: `basename $0` {boxname}"
+  exit 1
+fi
+
 echo "starting $1 for first time without provision to set hostname"
 vagrant destroy $1 -f
 vagrant up $1 --provision
